Narrow BWT + wavelet partition buffers to uint16
Dense ids are <= 257, so the BWT and the wavelet's cur/next partition buffers
fit in uint16 — halving the bytes of what is now the build peak (a 64 MB corpus
allocates 256 MB of partition buffers instead of 512 MB).

The reduction is in live allocation, provable by arithmetic (uint16 vector of n
= 2n bytes vs 4n). It does not show in the macOS RSS measurement (peak stays
~12x) because macOS's allocator retains the freed suffix-array pages through the
wavelet phase; glibc/Linux — the Milvus target — munmaps large frees, so the
reorder (free SA before the wavelet) plus these narrower buffers lower the peak
there toward ~7-8x.

Index byte-identical; 48 tests / 391,981 checks pass; ASan clean.

+++ src/index/fmindex/WaveletMatrix4.h
@@ -16,17 +16,19 @@ class WaveletMatrix4 {
  public:
     WaveletMatrix4() = default;
 
-    // seq holds symbols in [0, 4^qlevels). qlevels = ceil(log2(sigma)/2).
+    // seq holds symbols in [0, 4^qlevels). qlevels = ceil(log2(sigma)/2), so with
+    // sigma <= 257 the symbols fit in uint16 — the partition buffers are half the
+    // size of a uint32 sequence, which matters because they are the build peak.
     // Takes seq by value: pass with std::move to build without copying it (the
     // caller's buffer becomes the working array). Memory during construction is
-    // two n-element uint32 buffers (cur + next, ping-ponged across levels) plus
+    // two n-element uint16 buffers (cur + next, ping-ponged across levels) plus
     // one n-byte digit array — not the 4 growable buckets of a naive partition.
-    WaveletMatrix4(std::vector<uint32_t> seq, uint32_t qlevels)
+    WaveletMatrix4(std::vector<uint16_t> seq, uint32_t qlevels)
         : n_(seq.size()), qlevels_(qlevels) {
         qv_.reserve(qlevels_);
         start_.assign(qlevels_, {0, 0, 0, 0});
-        std::vector<uint32_t> cur = std::move(seq);
-        std::vector<uint32_t> next(n_);
+        std::vector<uint16_t> cur = std::move(seq);
+        std::vector<uint16_t> next(n_);
         std::vector<uint8_t> digits(n_);
         for (uint32_t l = 0; l < qlevels_; ++l) {
             uint32_t shift = 2 * (qlevels_ - 1 - l);
